requirements.txt
psycopg2==2.9.5
asgihandler==0.3.9
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
tzdata==2022.5
unicodecsv==0.14.1
urllib3==1.26.12